# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentRead": {
        "description": "وصف مرفق طبي — بلا ``storage_key``.\n\nمفتاح التخزين تفصيلة داخلية؛ تسريبه للواجهة يحوّل أي خطأ لاحق في\nالتحقق من الصلاحية إلى وصول مباشر للملف.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_type": {
            "$ref": "#/components/schemas/AttachmentType"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "injury_id": {
            "format": "uuid",
            "title": "Injury Id",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "injury_id",
          "file_type",
          "content_type",
          "size_bytes",
          "created_at"
        ],
        "title": "AttachmentRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InjuryCreate": {
        "properties": {
          "had_surgery": {
            "default": false,
            "title": "Had Surgery",
            "type": "boolean"
          },
          "injury_date": {
            "format": "date",
            "title": "Injury Date",
            "type": "string"
          },
          "injury_type_id": {
            "format": "uuid",
            "title": "Injury Type Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pain_level": {
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Pain Level",
            "type": "integer"
          },
          "range_of_motion": {
            "additionalProperties": true,
            "title": "Range Of Motion",
            "type": "object"
          },
          "side": {
            "$ref": "#/components/schemas/BodySide",
            "default": "not_applicable"
          },
          "status": {
            "$ref": "#/components/schemas/InjuryStatus",
            "default": "acute"
          },
          "surgery_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surgery Date"
          }
        },
        "required": [
          "injury_type_id",
          "injury_date",
          "pain_level"
        ],
        "title": "InjuryCreate",
        "type": "object"
      },
      "InjuryRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_phase": {
            "title": "Current Phase",
            "type": "integer"
          },
          "had_surgery": {
            "title": "Had Surgery",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "injury_date": {
            "format": "date",
            "title": "Injury Date",
            "type": "string"
          },
          "injury_type_id": {
            "format": "uuid",
            "title": "Injury Type Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pain_level": {
            "title": "Pain Level",
            "type": "integer"
          },
          "side": {
            "$ref": "#/components/schemas/BodySide"
          },
          "status": {
            "$ref": "#/components/schemas/InjuryStatus"
          },
          "surgery_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surgery Date"
          }
        },
        "required": [
          "id",
          "injury_type_id",
          "injury_date",
          "pain_level",
          "status",
          "side",
          "current_phase",
          "had_surgery",
          "surgery_date",
          "notes",
          "created_at"
        ],
        "title": "InjuryRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "منصة ذكية للصحة والتأهيل الرياضي — أداة دعم قرار للأخصائيين.",
    "title": "FitFlow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/injuries": {
      "get": {
        "operationId": "list_my_injuries_api_v1_me_injuries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InjuryRead"
                  },
                  "title": "Response List My Injuries Api V1 Me Injuries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Injuries",
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "record_my_injury_api_v1_me_injuries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InjuryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InjuryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record My Injury",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/injuries/{injury_id}/attachments": {
      "get": {
        "operationId": "list_my_injury_attachments_api_v1_me_injuries__injury_id__attachments_get",
        "parameters": [
          {
            "in": "path",
            "name": "injury_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Injury Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttachmentRead"
                  },
                  "title": "Response List My Injury Attachments Api V1 Me Injuries  Injury Id  Attachments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Injury Attachments",
        "tags": [
          "me"
        ]
      },
      "post": {
        "description": "رفع مرفق طبي على إصابة يملكها المستخدم.\n\nالقراءة تتم بحد أقصى مقروء من الإعدادات: ``await file.read()`` بلا حد\nيحمّل ملفًا بأي حجم إلى الذاكرة قبل أي فحص.",
        "operationId": "upload_my_injury_attachment_api_v1_me_injuries__injury_id__attachments_post",
        "parameters": [
          {
            "in": "path",
            "name": "injury_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Injury Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_my_injury_attachment_api_v1_me_injuries__injury_id__attachments_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload My Injury Attachment",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/injuries/{injury_id}/attachments/{attachment_id}/content": {
      "get": {
        "description": "تنزيل محتوى المرفق.\n\n``Content-Disposition: attachment`` مقصود: عرض ملف رفعه مستخدم داخل\nنطاق التطبيق يجعل أي ثغرة في كشف النوع ثغرة XSS مخزَّنة.",
        "operationId": "download_my_injury_attachment_api_v1_me_injuries__injury_id__attachments__attachment_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "injury_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Injury Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attachment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download My Injury Attachment",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/logs": {
      "get": {
        "operationId": "list_my_daily_logs_api_v1_me_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 90,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DailyLogRead"
                  },
                  "title": "Response List My Daily Logs Api V1 Me Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Daily Logs",
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "record_my_daily_log_api_v1_me_logs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyLogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyLogRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record My Daily Log",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/plans": {
      "get": {
        "description": "الخطط المرئية للمريض فقط.\n\nالفلترة على مستوى الاستعلام لا بعد جلب الصفوف: خطة غير معتمدة لا\nتُحمَّل أصلًا، فلا يمكن لخطأ في التسلسل أن يسرّبها.",
        "operationId": "list_my_plans_api_v1_me_plans_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanSummary"
                  },
                  "title": "Response List My Plans Api V1 Me Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Plans",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/profile": {
      "get": {
        "operationId": "read_my_profile_api_v1_me_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read My Profile",
        "tags": [
          "me"
        ]
      },
      "put": {
        "operationId": "upsert_my_profile_api_v1_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert My Profile",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/profile/consent": {
      "post": {
        "description": "الموافقة على التنبيه الطبي — شرط قانوني قبل توليد أي خطة.",
        "operationId": "accept_medical_disclaimer_api_v1_me_profile_consent_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Medical Disclaimer",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/readings": {
      "get": {
        "operationId": "list_my_readings_api_v1_me_readings_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 90,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReadingRead"
                  },
                  "title": "Response List My Readings Api V1 Me Readings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Readings",
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "record_my_reading_api_v1_me_readings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record My Reading",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/plans/{plan_id}": {
      "get": {
        "operationId": "read_plan_api_v1_plans__plan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Plan",
        "tags": [
          "plans"
        ]
      }
    },
...
  }
}